requirements.txt
moviepy==2.0.0.dev2
instagrapi
python-dotenv
pydantic
